Shift timeout errors to calling processes
While not a long term solution its impossible to tell currently if this
is just kube/gRPC being slow or if there's some other chicanery going on
in here. For now I'm going to shift the timeout errors further up the
call chain to see if these will complete given more time.