Fix double response bug
I was missing an early return after checking the HTTP verb. This lead to
accidentally queueing two response messages which violates an assertion
in boost that dictates only one outstanding read and write operation is
valid. And beyond that, it would also have broken the HTTP message
protocol as well had that assertion not triggered.

The bug is a bit racy. To increase the chances of triggeirng it you can
specify the same URL multiple times in a cURL call:

    $ curl http://127.0.0.1:50051/Ateles/Execute \
            http://127.0.0.1:50051/Ateles/Execute

Thanks to @nickva for reporting the issue.